Moved ckeditor component from lib to src/ckeditor.
