+++ src/ckeditor/ckeditor.component.spec.ts
