+++ src/ckeditor/ckeditor.component.ts
