+++ src/ckeditor/ckeditor.module.ts
