+++ src/ckeditor/package.json
@@ -22,6 +22,6 @@
     "whitelistedNonPeerDependencies": [
       "."
     ],
-    "dest": "../dist"
+    "dest": "../../dist"
   }
 }
