+++ src/test.ts
@@ -18,9 +18,5 @@ getTestBed().initTestEnvironment(
 // Then we find all the tests.
 const context = require.context( '.', true, /\.spec\.ts$/ );
 
-const libTestFiles = require.context( '../lib', true, /\.spec\.ts$/ );
-
 // And load the modules.
 context.keys().map( context );
-
-libTestFiles.keys().map( libTestFiles );
